requirements_freeze.txt
# Don't edit. This file is generated by poetry_update_deps.sh
-e .
allure-pytest==2.13.5 ; python_version >= "3.9" and python_version < "4.0"
allure-python-commons==2.13.5 ; python_version >= "3.9" and python_version < "4.0"
annotated-types==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.9" and python_version < "4.0"
cardano-clusterlib==0.6.0a19 ; python_version >= "3.9" and python_version < "4.0"
cbor2==5.6.3 ; python_version >= "3.9" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.9" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
cryptography==42.0.5 ; python_version >= "3.9" and python_version < "4.0"
deprecated==1.2.14 ; python_version >= "3.9" and python_version < "4.0"
exceptiongroup==1.2.1 ; python_version >= "3.9" and python_version < "3.11"
execnet==2.1.1 ; python_version >= "3.9" and python_version < "4.0"
filelock==3.14.0 ; python_version >= "3.9" and python_version < "4.0"
hypothesis==6.100.2 ; python_version >= "3.9" and python_version < "4.0"
idna==3.7 ; python_version >= "3.9" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.9" and python_version < "4.0"
packaging==24.0 ; python_version >= "3.9" and python_version < "4.0"
pluggy==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.9" and python_version < "4.0"
py==1.11.0 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.9" and python_version < "4.0"
pydantic-core==2.18.2 ; python_version >= "3.9" and python_version < "4.0"
pydantic==2.7.1 ; python_version >= "3.9" and python_version < "4.0"
pygithub==1.59.1 ; python_version >= "3.9" and python_version < "4.0"
pyjwt[crypto]==2.8.0 ; python_version >= "3.9" and python_version < "4.0"
pynacl==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
pytest-html==3.2.0 ; python_version >= "3.9" and python_version < "4.0"
pytest-metadata==3.1.1 ; python_version >= "3.9" and python_version < "4.0"
pytest-order==1.2.1 ; python_version >= "3.9" and python_version < "4.0"
pytest-select==0.1.2 ; python_version >= "3.9" and python_version < "4.0"
pytest-xdist==3.3.1 ; python_version >= "3.9" and python_version < "4.0"
pytest==7.4.4 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.9" and python_version < "4.0"
requests==2.32.2 ; python_version >= "3.9" and python_version < "4.0"
sortedcontainers==2.4.0 ; python_version >= "3.9" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.9" and python_version < "3.11"
typing-extensions==4.11.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.9" and python_version < "4.0"
wrapt==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
